Fix triggers to avoid infinite loops

+++ otte_docker_deployment/db_creation_scripts/db_functions_colony_assets.sql
@@ -1,4 +1,3 @@
-
 -- Triggers:
 
 ---------------------------------- Update AssetCollection Entries ----------------------------------
@@ -6,18 +5,21 @@
 CREATE OR REPLACE FUNCTION update_assetcollection_entries()
 RETURNS TRIGGER AS $$
 BEGIN
-    IF NEW."assetCollection" IS NOT NULL THEN
-        -- Case 1: If the assetCollection is not null, append the new id to the entries array
-        UPDATE "AssetCollection"
-        SET "collectionEntries" = array_append("collectionEntries", NEW.id)
-        WHERE id = NEW."assetCollection";
-    END IF;
-    
-    IF OLD."assetCollection" IS NOT NULL AND NEW."assetCollection" IS DISTINCT FROM OLD."assetCollection" THEN
-        -- Case 2: If the old assetCollection is not null and differs from the new one, remove the id from the old collection
-        UPDATE "AssetCollection"
-        SET "collectionEntries" = array_remove("collectionEntries", OLD.id)
-        WHERE id = OLD."assetCollection";
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        IF NEW."assetCollection" IS NOT NULL THEN
+            -- Case 1: If the assetCollection is not null, append the new id to the entries array
+            UPDATE "AssetCollection"
+            SET "collectionEntries" = array_append("collectionEntries", NEW.id)
+            WHERE id = NEW."assetCollection";
+        END IF;
+        
+        IF OLD."assetCollection" IS NOT NULL AND NEW."assetCollection" IS DISTINCT FROM OLD."assetCollection" THEN
+            -- Case 2: If the old assetCollection is not null and differs from the new one, remove the id from the old collection
+            UPDATE "AssetCollection"
+            SET "collectionEntries" = array_remove("collectionEntries", OLD.id)
+            WHERE id = OLD."assetCollection";
+        END IF;
     END IF;
     
     RETURN NEW;
@@ -34,11 +36,14 @@ EXECUTE FUNCTION update_assetcollection_entries();
 CREATE OR REPLACE FUNCTION delete_collectionentry_trigger()
 RETURNS TRIGGER AS $$
 BEGIN
-    IF OLD."assetCollection" IS NOT NULL THEN
-        -- Remove the entry from the array when the entry is deleted
-        UPDATE "AssetCollection"
-        SET "collectionEntries" = array_remove("collectionEntries", OLD.id)
-        WHERE id = OLD."assetCollection";
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        IF OLD."assetCollection" IS NOT NULL THEN
+            -- Remove the entry from the array when the entry is deleted
+            UPDATE "AssetCollection"
+            SET "collectionEntries" = array_remove("collectionEntries", OLD.id)
+            WHERE id = OLD."assetCollection";
+        END IF;
     END IF;
     
     RETURN OLD;
@@ -55,10 +60,13 @@ EXECUTE FUNCTION delete_collectionentry_trigger();
 CREATE OR REPLACE FUNCTION sync_collectionentry_addition()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Update the CollectionEntry to point to the AssetCollection
-    UPDATE "CollectionEntry"
-    SET "assetCollection" = NEW.id
-    WHERE id = ANY(NEW."collectionEntries");
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Update the CollectionEntry to point to the AssetCollection
+        UPDATE "CollectionEntry"
+        SET "assetCollection" = NEW.id
+        WHERE id = ANY(NEW."collectionEntries");
+    END IF;
     
     RETURN NEW;
 END;
@@ -73,10 +81,13 @@ EXECUTE FUNCTION sync_collectionentry_addition();
 CREATE OR REPLACE FUNCTION sync_collectionentry_removal()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Set the CollectionEntry assetCollection field to NULL if the ID is removed
-    DELETE FROM "CollectionEntry" 
-    WHERE id = ANY(OLD."collectionEntries")
-      AND id != ALL(NEW."collectionEntries");
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Set the CollectionEntry assetCollection field to NULL if the ID is removed
+        DELETE FROM "CollectionEntry" 
+        WHERE id = ANY(OLD."collectionEntries")
+          AND id != ALL(NEW."collectionEntries");
+    END IF;
     
     RETURN NEW;
 END;
@@ -93,20 +104,23 @@ EXECUTE FUNCTION sync_collectionentry_removal();
 CREATE OR REPLACE FUNCTION update_colony_locations()
 RETURNS TRIGGER AS $$
 BEGIN
-    IF NEW.colony IS NOT NULL THEN
-        -- Case 1: Add the `ColonyLocation` ID to the `locations` array in `Colony`
-        UPDATE "Colony"
-        SET locations = array_append(locations, NEW.id)
-        WHERE id = NEW.colony;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        IF NEW.colony IS NOT NULL THEN
+            -- Case 1: Add the `ColonyLocation` ID to the `locations` array in `Colony`
+            UPDATE "Colony"
+            SET locations = array_append(locations, NEW.id)
+            WHERE id = NEW.colony;
+        END IF;
+        
+        IF OLD.colony IS NOT NULL AND NEW.colony IS DISTINCT FROM OLD.colony THEN
+            -- Case 2: Remove the `ColonyLocation` ID from the old `Colony`
+            UPDATE Colony
+            SET locations = array_remove(locations, OLD.id)
+            WHERE id = OLD.colony;
+        END IF;
     END IF;
-    
-    IF OLD.colony IS NOT NULL AND NEW.colony IS DISTINCT FROM OLD.colony THEN
-        -- Case 2: Remove the `ColonyLocation` ID from the old `Colony`
-        UPDATE Colony
-        SET locations = array_remove(locations, OLD.id)
-        WHERE id = OLD.colony;
-    END IF;
-    
+
     RETURN NEW;
 END;
 $$ LANGUAGE plpgsql;
@@ -119,11 +133,14 @@ EXECUTE FUNCTION update_colony_locations();
 CREATE OR REPLACE FUNCTION delete_colonylocation_trigger()
 RETURNS TRIGGER AS $$
 BEGIN
-    IF OLD.colony IS NOT NULL THEN
-        -- Remove the `ColonyLocation` ID from the `locations` array in `Colony`
-        UPDATE "Colony"
-        SET locations = array_remove(locations, OLD.id)
-        WHERE id = OLD.colony;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        IF OLD.colony IS NOT NULL THEN
+            -- Remove the `ColonyLocation` ID from the `locations` array in `Colony`
+            UPDATE "Colony"
+            SET locations = array_remove(locations, OLD.id)
+            WHERE id = OLD.colony;
+        END IF;
     END IF;
     
     RETURN OLD;
@@ -140,11 +157,14 @@ EXECUTE FUNCTION delete_colonylocation_trigger();
 CREATE OR REPLACE FUNCTION sync_colonylocation_addition()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Update `ColonyLocation` to point to the `Colony`
-    UPDATE "ColonyLocation"
-    SET colony = NEW.id
-    WHERE id = ANY(NEW.locations);
-    
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Update `ColonyLocation` to point to the `Colony`
+        UPDATE "ColonyLocation"
+        SET colony = NEW.id
+        WHERE id = ANY(NEW.locations);
+    END IF;
+
     RETURN NEW;
 END;
 $$ LANGUAGE plpgsql;
@@ -158,12 +178,15 @@ EXECUTE FUNCTION sync_colonylocation_addition();
 CREATE OR REPLACE FUNCTION sync_colonylocation_removal()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Set the `ColonyLocation` `colony` field to NULL if the ID is removed
-    UPDATE "ColonyLocation"
-    SET colony = NULL
-    WHERE id = ANY(OLD.locations)
-      AND id != ALL(NEW.locations);
-    
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Set the `ColonyLocation` `colony` field to NULL if the ID is removed
+        UPDATE "ColonyLocation"
+        SET colony = NULL
+        WHERE id = ANY(OLD.locations)
+          AND id != ALL(NEW.locations);
+    END IF;
+
     RETURN NEW;
 END;
 $$ LANGUAGE plpgsql;
@@ -179,20 +202,23 @@ EXECUTE FUNCTION sync_colonylocation_removal();
 CREATE OR REPLACE FUNCTION update_colony_assets()
 RETURNS TRIGGER AS $$
 BEGIN
-    IF NEW.colony IS NOT NULL THEN
-        -- Case 1: Add the `ColonyAsset` ID to the `assets` array in `Colony`
-        UPDATE "Colony"
-        SET assets = array_append(assets, NEW.id)
-        WHERE id = NEW.colony;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        IF NEW.colony IS NOT NULL THEN
+            -- Case 1: Add the `ColonyAsset` ID to the `assets` array in `Colony`
+            UPDATE "Colony"
+            SET assets = array_append(assets, NEW.id)
+            WHERE id = NEW.colony;
+        END IF;
+        
+        IF OLD.colony IS NOT NULL AND NEW.colony IS DISTINCT FROM OLD.colony THEN
+            -- Case 2: Remove the `ColonyAsset` ID from the old `Colony`
+            UPDATE "Colony"
+            SET assets = array_remove(assets, OLD.id)
+            WHERE id = OLD.colony;
+        END IF;
     END IF;
-    
-    IF OLD.colony IS NOT NULL AND NEW.colony IS DISTINCT FROM OLD.colony THEN
-        -- Case 2: Remove the `ColonyAsset` ID from the old `Colony`
-        UPDATE "Colony"
-        SET assets = array_remove(assets, OLD.id)
-        WHERE id = OLD.colony;
-    END IF;
-    
+
     RETURN NEW;
 END;
 $$ LANGUAGE plpgsql;
@@ -205,11 +231,14 @@ EXECUTE FUNCTION update_colony_assets();
 CREATE OR REPLACE FUNCTION delete_colonyasset_trigger()
 RETURNS TRIGGER AS $$
 BEGIN
-    IF OLD.colony IS NOT NULL THEN
-        -- Remove the `ColonyAsset` ID from the `assets` array in `Colony`
-        UPDATE "Colony"
-        SET assets = array_remove(assets, OLD.id)
-        WHERE id = OLD.colony;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        IF OLD.colony IS NOT NULL THEN
+            -- Remove the `ColonyAsset` ID from the `assets` array in `Colony`
+            UPDATE "Colony"
+            SET assets = array_remove(assets, OLD.id)
+            WHERE id = OLD.colony;
+        END IF;
     END IF;
     
     RETURN OLD;
@@ -226,10 +255,13 @@ EXECUTE FUNCTION delete_colonyasset_trigger();
 CREATE OR REPLACE FUNCTION sync_colonyasset_addition()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Update `ColonyAsset` to point to the `Colony`
-    UPDATE "ColonyAsset"
-    SET colony = NEW.id
-    WHERE id = ANY(NEW.assets);
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Update `ColonyAsset` to point to the `Colony`
+        UPDATE "ColonyAsset"
+        SET colony = NEW.id
+        WHERE id = ANY(NEW.assets);
+    END IF;
     
     RETURN NEW;
 END;
@@ -244,11 +276,14 @@ EXECUTE FUNCTION sync_colonyasset_addition();
 CREATE OR REPLACE FUNCTION sync_colonyasset_removal()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Set the `ColonyAsset` `colony` field to NULL if the ID is removed
-    UPDATE "ColonyAsset"
-    SET colony = NULL
-    WHERE id = ANY(OLD.assets)
-      AND id != ALL(NEW.assets);
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Set the `ColonyAsset` `colony` field to NULL if the ID is removed
+        UPDATE "ColonyAsset"
+        SET colony = NULL
+        WHERE id = ANY(OLD.assets)
+          AND id != ALL(NEW.assets);
+    END IF;
     
     RETURN NEW;
 END;
@@ -265,10 +300,13 @@ EXECUTE FUNCTION sync_colonyasset_removal();
 CREATE OR REPLACE FUNCTION sync_colony_update()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Sync ColonyCode's colony when Colony is updated
-    UPDATE "ColonyCode"
-    SET colony = NEW.id
-    WHERE id = NEW."colonyCode";
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Sync ColonyCode's colony when Colony is updated
+        UPDATE "ColonyCode"
+        SET colony = NEW.id
+        WHERE id = NEW."colonyCode";
+    END IF;
 
     RETURN NEW;
 END;
@@ -282,10 +320,13 @@ EXECUTE FUNCTION sync_colony_update();
 CREATE OR REPLACE FUNCTION handle_colony_deletion()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Set ColonyCode's colony to NULL when corresponding Colony is deleted
-    UPDATE "ColonyCode"
-    SET colony = NULL
-    WHERE colony = OLD.id;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Set ColonyCode's colony to NULL when corresponding Colony is deleted
+        UPDATE "ColonyCode"
+        SET colony = NULL
+        WHERE colony = OLD.id;
+    END IF;
 
     RETURN OLD;
 END;
@@ -301,10 +342,13 @@ EXECUTE FUNCTION handle_colony_deletion();
 CREATE OR REPLACE FUNCTION sync_colonycode_update()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Sync Colony's colonyCode when ColonyCode is updated
-    UPDATE "Colony"
-    SET "colonyCode" = NEW.id
-    WHERE id = NEW.colony;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Sync Colony's colonyCode when ColonyCode is updated
+        UPDATE "Colony"
+        SET "colonyCode" = NEW.id
+        WHERE id = NEW.colony;
+    END IF;
 
     RETURN NEW;
 END;
@@ -318,10 +362,13 @@ EXECUTE FUNCTION sync_colonycode_update();
 CREATE OR REPLACE FUNCTION handle_colonycode_deletion()
 RETURNS TRIGGER AS $$
 BEGIN
-    -- Set Colony's colonyCode to NULL when corresponding ColonyCode is deleted
-    UPDATE "Colony"
-    SET "colonyCode" = NULL
-    WHERE "colonyCode" = OLD.id;
+    -- Check the trigger depth
+    IF pg_trigger_depth() = 0 THEN
+        -- Set Colony's colonyCode to NULL when corresponding ColonyCode is deleted
+        UPDATE "Colony"
+        SET "colonyCode" = NULL
+        WHERE "colonyCode" = OLD.id;
+    END IF;
 
     RETURN OLD;
 END;
@@ -332,7 +379,6 @@ AFTER DELETE ON "ColonyCode"
 FOR EACH ROW
 EXECUTE FUNCTION handle_colonycode_deletion();
 ---------------------------------- Update Colony ----------------------------------
-
 
 
 -- Functions:
@@ -352,5 +398,3 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 ---------------------------------- getLocationApperances ----------------------------------
-
-
